+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Changed
+- Renamed folder 'images' to 'resources' in projects directory.
+
 ### Fixed
 - Command now keeps prompting for the folder name when it's invalid.