+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+
+
+## [1.4.0] – 2025-08-21
 ### Added
 - Validation for structure names:
   - Shows modal error message when an empty structure name is detected.